- Some more prep for SIMD on s390x (mono/mono#16903)
- Use __builtin_bswapxx for s390x rather than inline assembler
- s390x ABI specifies that %f8-%f15 are protected between calls so avoid using or save/restore as required
- Save the "protected" FP registers in the context structure

Commit migrated from https://github.com/mono/mono/commit/86f7cef7f43f9289e9aea96799875ed9578c50ab

+++ src/mono/mono/mini/exceptions-s390x.c
@@ -57,6 +57,7 @@
 #include <mono/metadata/debug-helpers.h>
 #include <mono/metadata/exception.h>
 #include <mono/metadata/mono-debug.h>
+#include <mono/utils/mono-hwcap.h>
 
 #include "mini.h"
 #include "mini-s390x.h"
@@ -339,6 +340,7 @@ mono_arch_get_throw_exception_generic (int size, MonoTrampInfo **info, int corli
 		s390_std (code, i, 0, STK_BASE, pos);
 		pos += sizeof (gdouble);
 	}
+
 	/*------------------------------------------------------*/
 	/* save the access registers         			*/
 	/*------------------------------------------------------*/
@@ -499,7 +501,7 @@ mono_arch_unwind_frame (MonoDomain *domain, MonoJitTlsData *jit_tls,
 	*new_ctx = *ctx;
 
 	if (ji != NULL) {
-		gint64 address;
+		uintptr_t address;
 		guint8 *cfa;
 		guint32 unwind_info_len;
 		guint8 *unwind_info;
@@ -633,7 +635,7 @@ mono_arch_setup_async_callback (MonoContext *ctx, void (*async_cb)(void *fun), g
 	ctx->uc_mcontext.gregs[2] = (gsize)user_data;
 
 	sp -= S390_MINIMAL_STACK_SIZE;
-	*(unsigned long *)sp = MONO_CONTEXT_GET_SP(ctx);
+	*(unsigned long *)sp = (uintptr_t) MONO_CONTEXT_GET_SP(ctx);
 	MONO_CONTEXT_SET_BP(ctx, sp);
 	MONO_CONTEXT_SET_IP(ctx, (gsize)async_cb);
 }
